app/requirements.txt
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
mixer==7.2.2
psycopg2-binary==2.9.4

